# app/api/statistics/products/route.ts
import { NextResponse } from 'next/server';
import { getProductStatistics } from '@/server/domain/statistics/statistics.service';
import { ZodError } from 'zod';

export async function GET(request: Request) {
  try {
    const { searchParams } = new URL(request.url);
    const query = Object.fromEntries(searchParams.entries());
    const statistics = await getProductStatistics(query);
    return NextResponse.json(statistics);
  } catch (error) {
    if (error instanceof ZodError) {
      return NextResponse.json({ message: 'Validation error', errors: error.errors }, { status: 400 });
    }
    console.error(error);
    return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
  }
}