# app/api/suppliers/route.ts
import { NextResponse } from 'next/server';
import { getSuppliers, createSupplier } from '@/server/domain/suppliers/suppliers.service';
import { ZodError } from 'zod';

export async function GET() {
  try {
    const suppliers = await getSuppliers();
    return NextResponse.json(suppliers);
  } catch (error) {
    console.error(error);
    return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
  }
}

export async function POST(request: Request) {
  try {
    const body = await request.json();
    const newSupplier = await createSupplier(body);
    return NextResponse.json(newSupplier, { status: 201 });
  } catch (error) {
    if (error instanceof ZodError) {
      return NextResponse.json({ message: 'Validation error', errors: error.errors }, { status: 400 });
    }
    console.error(error);
    return NextResponse.json({ message: 'Internal server error' }, { status: 500 });
  }
}