# public/openapi.yaml
openapi: 3.0.3
info:
  title: Accessory Management API
  version: 0.2.0
  description: >
    프론트엔드 목업 기반으로 작성한 REST API 명세.
    규칙(가정):
    - 인증: JWT Bearer (Authorization: Bearer <token>)
    - ID 규칙: 모든 리소스 id는 문자열(예: uuid 또는 shortid)
    - 페이징: list API는 page, pageSize(기본 20, 최대 100) 지원
    - 정렬: sortBy, sortDir(asc|desc) 지원(해당 리소스에서만)
    - 에러 응답: 공통 ErrorResponse 사용
servers:
  - url: https://api.example.com

security:
  - bearerAuth: []

tags:
  - name: Auth
  - name: Suppliers
  - name: Categories
  - name: Products
  - name: Stock
  - name: Orders
  - name: Statistics
  - name: Metadata

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: 인증 실패
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /suppliers:
    get:
      tags: [Suppliers]
      summary: 거래처 목록
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Supplier" }
    post:
      tags: [Suppliers]
      summary: 거래처 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierCreate" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }

  /suppliers/{supplierId}:
    patch:
      tags: [Suppliers]
      summary: 거래처 수정
      parameters:
        - name: supplierId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierUpdate" }
      responses:
        "200":
          description: 수정됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
    delete:
      tags: [Suppliers]
      summary: 거래처 삭제
      parameters:
        - name: supplierId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 삭제됨

  /categories:
    get:
      tags: [Categories]
      summary: 카테고리 목록
      responses:
        "200":
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }

  /products:
    get:
      tags: [Products]
      summary: 상품 목록
      parameters:
        - name: supplierId
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/sortDir"
      responses:
        "200":
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      tags: [Products]
      summary: 상품 단건 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductCreate" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }

  /products/bulk:
    post:
      tags: [Products]
      summary: 상품 일괄 생성 (이미지 포함)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductBulkCreate" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }

  /products/{productId}:
    get:
      tags: [Products]
      summary: 상품 상세
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    patch:
      tags: [Products]
      summary: 상품 수정
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductUpdate" }
      responses:
        "200":
          description: 수정됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      tags: [Products]
      summary: 상품 삭제
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 삭제됨

  /products/{productId}/stock:
    get:
      tags: [Stock]
      summary: 상세 재고 조회
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 상세 재고
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockDetail" }
    put:
      tags: [Stock]
      summary: 상세 재고 저장
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockUpdate" }
      responses:
        "200":
          description: 저장됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockDetail" }

  /orders:
    get:
      tags: [Orders]
      summary: 발주 목록
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/OrderStatus" }
        - name: supplierId
          in: query
          schema: { type: string }
        - name: includeItems
          in: query
          schema: { type: boolean, default: false }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      tags: [Orders]
      summary: 발주 생성 (전송)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreate" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: 발주 상세
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
    patch:
      tags: [Orders]
      summary: 발주 상태 변경
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderUpdate" }
      responses:
        "200":
          description: 변경됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }

  /statistics/products:
    get:
      tags: [Statistics]
      summary: 기간/필터 기반 품목별 통계
      parameters:
        - name: startYear
          in: query
          required: true
          schema: { type: string, example: "2023" }
        - name: startMonth
          in: query
          required: true
          schema: { type: string, example: "10" }
        - name: endYear
          in: query
          required: true
          schema: { type: string, example: "2023" }
        - name: endMonth
          in: query
          required: true
          schema: { type: string, example: "11" }
        - name: supplierId
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: color
          in: query
          schema: { $ref: "#/components/schemas/Color" }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [quantity, amount]
            default: quantity
      responses:
        "200":
          description: 통계 결과
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductStatisticsResponse" }

  /metadata:
    get:
      tags: [Metadata]
      summary: 메타데이터(색상/호수/상태 등)
      responses:
        "200":
          description: 메타데이터
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Metadata" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    sortBy:
      name: sortBy
      in: query
      schema: { type: string }
    sortDir:
      name: sortDir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "INVALID_REQUEST" }
        message: { type: string, example: "요청이 올바르지 않습니다." }
        details:
          type: object
          additionalProperties: true

    LoginRequest:
      type: object
      required: [id, password]
      properties:
        id: { type: string }
        password: { type: string }

    LoginResponse:
      type: object
      required: [accessToken, tokenType]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, example: "Bearer" }
        expiresIn: { type: integer, example: 3600 }
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }

    Supplier:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }

    SupplierCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }

    SupplierUpdate:
      type: object
      properties:
        name: { type: string }

    Category:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        count: { type: integer }
        icon: { type: string }

    Color:
      type: string
      enum: ["골드", "로즈", "실버"]

    RingSize:
      type: string
      enum: ["3호","5호","7호","9호","11호","13호","15호","17호","19호","21호","23호"]

    OrderStatus:
      type: string
      enum: ["대기", "완료"]

    Product:
      type: object
      required: [id, name, sku, price, category, imageUrl, supplierId, stock]
      properties:
        id: { type: string }
        name: { type: string }
        sku: { type: string }
        price: { type: number, example: 120000 }
        category: { type: string }
        imageUrl: { type: string }
        supplierId: { type: string }
        supplierName: { type: string }
        stock: { type: integer }
        hasSizes: { type: boolean }

    ProductCreate:
      type: object
      required: [name, sku, price, category, supplierId, imageUrl]
      properties:
        name: { type: string }
        sku: { type: string }
        price: { type: number }
        category: { type: string }
        supplierId: { type: string }
        imageUrl: { type: string }
        hasSizes: { type: boolean }

    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        sku: { type: string }
        price: { type: number }
        category: { type: string }
        supplierId: { type: string }
        imageUrl: { type: string }
        hasSizes: { type: boolean }
        stock: { type: integer }

    ProductBulkCreate:
      type: object
      required: [supplierId, items]
      properties:
        supplierId: { type: string }
        items:
          type: array
          items:
            type: object
            required: [category, price]
            properties:
              category: { type: string }
              price: { type: number }
              imageBase64: { type: string, description: "data URL 또는 base64" }
              imageUrl: { type: string }

    OrderItem:
      type: object
      required: [productId, name, sku, price, category, supplierId, quantity, selectedColor]
      properties:
        productId: { type: string }
        name: { type: string }
        sku: { type: string }
        price: { type: number }
        category: { type: string }
        imageUrl: { type: string }
        supplierId: { type: string }
        supplierName: { type: string }
        quantity: { type: integer }
        selectedColor: { $ref: "#/components/schemas/Color" }
        selectedSize: { $ref: "#/components/schemas/RingSize" }

    Order:
      type: object
      required: [id, orderNumber, date, supplierId, itemCount, totalAmount, status]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        date: { type: string, example: "2023.10.24 • 오후 2:30" }
        supplierId: { type: string }
        supplierName: { type: string }
        itemCount: { type: integer }
        totalAmount: { type: number }
        status: { $ref: "#/components/schemas/OrderStatus" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }

    OrderCreate:
      type: object
      required: [supplierId, items]
      properties:
        supplierId: { type: string }
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, selectedColor]
            properties:
              productId: { type: string }
              quantity: { type: integer }
              selectedColor: { $ref: "#/components/schemas/Color" }
              selectedSize: { $ref: "#/components/schemas/RingSize" }

    OrderUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/OrderStatus" }

    StockVariant:
      type: object
      required: [color, quantity]
      properties:
        color: { $ref: "#/components/schemas/Color" }
        size: { $ref: "#/components/schemas/RingSize" }
        quantity: { type: integer }

    StockDetail:
      type: object
      required: [productId, total, variants]
      properties:
        productId: { type: string }
        total: { type: integer }
        variants:
          type: array
          items: { $ref: "#/components/schemas/StockVariant" }

    StockUpdate:
      type: object
      required: [variants]
      properties:
        variants:
          type: array
          items: { $ref: "#/components/schemas/StockVariant" }

    ProductStat:
      type: object
      required: [id, name, category, supplierId, totalQty, unitPrice, totalAmount]
      properties:
        id: { type: string }
        name: { type: string }
        imageUrl: { type: string }
        category: { type: string }
        supplierId: { type: string }
        supplierName: { type: string }
        totalQty: { type: integer }
        unitPrice: { type: number }
        totalAmount: { type: number }

    ProductStatisticsResponse:
      type: object
      properties:
        range:
          type: object
          properties:
            start: { type: string, example: "2023.10" }
            end: { type: string, example: "2023.11" }
        filters:
          type: object
          properties:
            supplierId: { type: string }
            category: { type: string }
            color: { $ref: "#/components/schemas/Color" }
            sortBy: { type: string, enum: [quantity, amount] }
        totals:
          type: object
          properties:
            qty: { type: integer }
            amount: { type: number }
        items:
          type: array
          items: { $ref: "#/components/schemas/ProductStat" }

    Metadata:
      type: object
      properties:
        colors:
          type: array
          items: { $ref: "#/components/schemas/Color" }
        ringSizes:
          type: array
          items: { $ref: "#/components/schemas/RingSize" }
        orderStatuses:
          type: array
          items: { $ref: "#/components/schemas/OrderStatus" }
        categories:
          type: array
          items: { $ref: "#/components/schemas/Category" }